Improve reconnection for IntesisBox
