+++ pyintesishome/__init__.py
@@ -1,3 +1,6 @@
 """ pyintesishome Module """
 from .exceptions import IHAuthenticationError, IHConnectionError
-from .pyintesishome import IntesisHome, IntesisHomeLocal, IntesisBox
+from .intesisbase import IntesisBase
+from .intesishome import IntesisHome
+from .intesishomelocal import IntesisHomeLocal
+from .intesisbox import IntesisBox
